+++ tests/test_metad.py
@@ -15,3 +15,20 @@ class TestMetaData(TestCase):
         metadata = MetaData.from_json(path_to_example)
         metadata.validate()
         assert str(metadata)
+
+    def test_creation(self):
+        metadata = MetaData()
+        metadata.add_table({
+            "id": "users",
+            "name": "users",
+            "fields": [
+                {"name": "user_id", "data_type": "id"},
+                {"name": "name", "data_type": "text"},
+                {"name": "gender", "data_type": "categorical"},
+                {"name": "age", "data_type": "numerical"},
+                {"name": "birthday", "data_type": "datetime"},
+                {"name": "picture", "data_type": "other"}
+            ]
+        })
+        assert len(metadata.get_table_names()) == 1
+        assert "users" in metadata.get_table_names()