+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## 0.0.1 (2020-06-08)
+
+* Populate data subtype in MySQL connector - [Issue #11](https://github.com/data-dev/MetaData/issues/11) by @k15z
+* Modifying the schema to support non-unique table names - [Issue #13](https://github.com/data-dev/MetaData/issues/13) by @k15z
+
 ## 0.0.1 (2020-05-22)
 
 * First release on PyPI.